add ProtocolException add ProtocolResponseDetector custom http header
