+++ shadowsocks-csharp/shadowsocks-csharp4.0.csproj
@@ -147,6 +147,7 @@
     <Compile Include="Controller\PACServer.cs" />
     <Compile Include="Model\Server.cs" />
     <Compile Include="Model\Configuration.cs" />
+    <Compile Include="Model\ServerSpeedLog.cs" />
     <Compile Include="Obfs\Auth.cs" />
     <Compile Include="Obfs\HttpSimpleObfs.cs" />
     <Compile Include="Obfs\IObfs.cs" />
